Move inner classes of BattleDisplay.java to their own class files
